+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
     - Dolfinx setup via docker for `real` and `compex` mode.
     - Repository environment package with  python requerements.
     - Loudspeaker config schema.
+    - FEM wave_equation solver.
+    - Signal generate and analyse tools.
 
 ---